requirements.txt
langchain
langchain-community
langchain-chroma
langchain-openai
pandas
faker
python-dotenv
sentence-transformers
